refactor camapign dir
