+++ src/pages/campaign/components/tooltip-enough-xrp.tsx
