+++ src/pages/campaign/pages/participate/hooks/use-step.ts
@@ -1,0 +1,3 @@
+// auto move
+// next prev
+// step status object
